AU-2329: feature: AU-2329: Refactored code.
