+++ e2e/utils/tooltip_validation_helpers.ts
@@ -1,8 +1,7 @@
 import {expect, Page, test} from "@playwright/test";
-import {FieldSwapItemList, FormData, FormPage, Selector, TooltipsList} from "./data/test_data";
+import {FormData, TooltipsList} from "./data/test_data";
 import {logger} from "./logger";
-import {clickButton, fillFormField} from "./input_helpers";
-import {logCurrentUrl} from "./helpers";
+import {goToSubmissionUrl, navigateToApplicationPage} from "./navigation_helpers";
 
 /**
  * The validateTooltips function.
@@ -79,47 +78,6 @@ const validateTooltipsOnPage = async (
     const tooltipContent = await page.locator('.tippy-content .webform-element-help--content').innerText();
     expect(tooltipContent.trim()).toBe(tooltip.message);
   }
-};
-
-/**
- * The goToSubmissionUrl function.
- *
- * This function navigates to an applications
- * submission URL.
- *
- * @param page
- *   Page object from Playwright.
- * @param submissionUrl
- *   A submission URL.
- */
-const goToSubmissionUrl = async (page: Page, submissionUrl: string) => {
-  await page.goto(submissionUrl);
-  await logCurrentUrl(page);
-  await page.waitForURL('**/muokkaa');
-  logger(`Navigated to: ${submissionUrl}.`);
-};
-
-/**
- * The navigateToApplicationPage function.
- *
- * This function navigate to a given page
- * (formPageKey) on an application.
- *
- * @param page
- *   Page object from Playwright.
- * @param formPageKey
- *   A form pages key (the page we are navigating to).
- */
-const navigateToApplicationPage = async (page: Page, formPageKey: string) => {
-  const applicantDetailsLink: Selector = {
-    type: 'form-topnavi-link',
-    name: 'data-webform-page',
-    value: formPageKey,
-  }
-  await clickButton(page, applicantDetailsLink);
-  await page.waitForLoadState('load');
-  await logCurrentUrl(page);
-  logger(`Loaded page: ${formPageKey}.`);
 };
 
 export { validateTooltips };
